+++ l10n_it_einvoice_send2sdi/egg-info/history.rst
@@ -1,3 +1,10 @@
+10.0.1.0.4 (2019-06-13)
+~~~~~~~~~~~~~~~~~~~~~~~
+
+* [FIX] Recoginze state "Notifica di esito: documento accettato"
+* [IMP] New button "Set to delivered" when recipient_error
+
+
 10.0.1.0.3 (2019-06-13)
 ~~~~~~~~~~~~~~~~~~~~~~~